Implement menu option 1
Adds a few static constants in Lab3binsearchTest to control the array
test size and value bounds.

Also adds a Binarysearch instance and int array to main. These will be
used be used by the other menu options when they are implemented.

The bounds parameters of generateRandomInts() and/or their Javadoc
comments should be renamed - this will be addressed in a future
commit. They currently do not match.

Similarly, there should be another newline after the array is printed.